Refactored SamlAssertionsVerifier
